Extracting more common code
