+++ packages/narada-core/src/narada_core/models.py
@@ -1,5 +1,7 @@
 from enum import Enum
-from typing import Literal, TypedDict
+from typing import Generic, Literal, TypedDict, TypeVar
+
+from pydantic import BaseModel
 
 
 class Agent(Enum):
@@ -22,3 +24,25 @@ class UserResourceCredentials(TypedDict, total=False):
 class RemoteDispatchChatHistoryItem(TypedDict):
     role: Literal["user", "assistant"]
     content: str
+
+
+_MaybeStructuredOutput = TypeVar("_MaybeStructuredOutput", bound=BaseModel | None)
+
+
+class ResponseContent(TypedDict, Generic[_MaybeStructuredOutput]):
+    text: str
+    structuredOutput: _MaybeStructuredOutput
+
+
+class Usage(TypedDict):
+    actions: int
+    credits: int
+
+
+class Response(TypedDict, Generic[_MaybeStructuredOutput]):
+    requestId: str
+    status: Literal["success", "error"]
+    response: ResponseContent[_MaybeStructuredOutput] | None
+    createdAt: str
+    completedAt: str | None
+    usage: Usage
