+++ packages/narada-pyodide/src/narada/__init__.py
@@ -1,4 +1,4 @@
-from narada_core.models import Agent
+from narada_core.models import Agent, Response, ResponseContent
 
 from narada.client import Narada
 from narada.errors import (
@@ -8,8 +8,6 @@ from narada.errors import (
 from narada.window import (
     LocalBrowserWindow,
     RemoteBrowserWindow,
-    Response,
-    ResponseContent,
 )
 
 __all__ = [
