+++ packages/narada-pyodide/src/narada/window.py
@@ -4,7 +4,7 @@ import os
 import time
 from abc import ABC
 from http import HTTPStatus
-from typing import TYPE_CHECKING, Any, Generic, Literal, TypedDict, TypeVar, overload
+from typing import TYPE_CHECKING, Any, Literal, TypeVar, overload
 
 from js import AbortController, setTimeout  # type: ignore
 from narada_core.actions.models import (
@@ -24,6 +24,7 @@ from narada_core.actions.models import (
 from narada_core.models import (
     Agent,
     RemoteDispatchChatHistoryItem,
+    Response,
     UserResourceCredentials,
 )
 from pydantic import BaseModel
@@ -38,28 +39,6 @@ if TYPE_CHECKING:
 
 
 _StructuredOutput = TypeVar("_StructuredOutput", bound=BaseModel)
-
-_MaybeStructuredOutput = TypeVar("_MaybeStructuredOutput", bound=BaseModel | None)
-
-
-class ResponseContent(TypedDict, Generic[_MaybeStructuredOutput]):
-    text: str
-    structuredOutput: _MaybeStructuredOutput
-
-
-class Usage(TypedDict):
-    actions: int
-    credits: int
-
-
-class Response(TypedDict, Generic[_MaybeStructuredOutput]):
-    requestId: str
-    status: Literal["success", "error"]
-    response: ResponseContent[_MaybeStructuredOutput] | None
-    createdAt: str
-    completedAt: str | None
-    usage: Usage
-
 
 _ResponseModel = TypeVar("_ResponseModel", bound=BaseModel)
